Move bootstrap code to bootstrap.cpp, make constructor for Actuator and LRTSpeedController explicit, fix comparison bug in ValueChangeEvent, make toggling output destination for AsyncPrinter isolated to AsyncPrinter.
Bootstrapping code has been moved to bootstrap.cpp to make location more explicit and obvious to find.  Constructors for Actuator and LRTSpeedController were previously not marked explicit, allowing for implicit creation of those classes and the potential for bugs.  A bugfix has been made in ValueChangeEvent.  Previously, the assignment operator was used instead of the equality operator, leading to the check to be assigning a value and checking that value rather than checking for equality.  Finally, the output destination toggle methods in AsyncPrinter have been modified so that they are isolated only to AsyncPrinter.